+++ src/app.ts
@@ -25,17 +25,17 @@ import * as promClient from 'prom-client';
 import swaggerUi from 'swagger-ui-express';
 import YAML from 'yaml';
 
-import { TrustedGatewayArNSResolver } from './arns/trusted-gateway-arns-resolver.js';
 import { ArweaveCompositeClient } from './arweave/composite-client.js';
 import { ReadThroughChunkDataCache } from './cache/read-through-chunk-data-cache.js';
 import { GatewayDataSource } from './data/gateway-data-source.js';
 import { SequentialDataSource } from './data/sequential-data-source.js';
-import { StreamingManifestPathResolver } from './data/streaming-manifest-path-resolver.js';
 import { TxChunksDataSource } from './data/tx-chunks-data-source.js';
 import { StandaloneSqliteDatabase } from './database/standalone-sqlite.js';
 import { UniformFailureSimulator } from './lib/chaos.js';
 import log from './log.js';
 import { createArnsMiddleware } from './middleware/arns.js';
+import { StreamingManifestPathResolver } from './resolvers/streaming-manifest-path-resolver.js';
+import { TrustedGatewayArNSResolver } from './resolvers/trusted-gateway-arns-resolver.js';
 import {
   DATA_PATH_REGEX,
   RAW_DATA_PATH_REGEX,
